The bubble scaling factor is now given as a double and not as an integer. Moved project to version 3.2.2.

+++ trunk/src/main/java/com/compomics/util/gui/spectrum/MassErrorBubblePlot.java
@@ -166,7 +166,7 @@ public class MassErrorBubblePlot extends JPanel {
             ArrayList<PeptideFragmentIon.PeptideFragmentIonType> currentFragmentIons,
             ArrayList<MSnSpectrum> currentSpectra,
             double massTolerance,
-            int bubbleScale,
+            double bubbleScale,
             boolean includeSinglyCharge,
             boolean includeDoublyCharge,
             boolean includeMoreThanTwoCharges,
